requirements.txt
opencv-python==4.13.0.92
Pillow==12.1.1
Adafruit-Blinka==8.70.0
adafruit-circuitpython-bme280==2.6.30
adafruit-circuitpython-bno08x==1.3.1
lgpio
